# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/account": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Créer un nouveau compte utilisateur",
                "operationId": "fa47c10a4036ffcd80cf08f3704f464d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securepassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Se connecter et obtenir un token JWT",
                "operationId": "796ba9db62ae42729918404eaff7d3f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie, token JWT renvoyé"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/carts": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get all items in the cart",
                "operationId": "cedd6bd3fdbb95998623732e9cb660b0",
                "responses": {
                    "200": {
                        "description": "List of cart items"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a product to the cart",
                "operationId": "c781f9f66f965a4efd238bcd1e99af56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added to cart"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/carts/{id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove a product from the cart",
                "operationId": "f990f9adb4898de626372cf361dae18c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from cart"
                    },
                    "404": {
                        "description": "Cart item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Liste des produits",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Liste des produits récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Créer un produit",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Afficher un produit spécifique",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du produit récupérés avec succès"
                    },
                    "404": {
                        "description": "Produit introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Supprimer un produit",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produit supprimé avec succès"
                    },
                    "404": {
                        "description": "Produit introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Mettre à jour un produit",
                "operationId": "f0caf1fe47821e6aad5d0c767b7476eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit mis à jour avec succès"
                    },
                    "404": {
                        "description": "Produit introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wishlists": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get all products in the wishlist",
                "operationId": "ac503bf2f1cfbca27567f5e6bb79a835",
                "responses": {
                    "200": {
                        "description": "List of wishlist items"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add a product to the wishlist",
                "operationId": "08914d855c8d22082e5d19cc047369e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added to wishlist"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wishlists/{id}": {
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Remove a product from the wishlist",
                "operationId": "2a89b9f79169b14ff7cfa19c6d3c0d30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from wishlist"
                    },
                    "404": {
                        "description": "Wishlist item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Cart": {
                "title": "Cart",
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "P001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Product Name"
                    },
                    "description": {
                        "type": "string",
                        "example": "Product description."
                    },
                    "image": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/image.jpg"
                    },
                    "category": {
                        "type": "string",
                        "example": "Category Name"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "internalReference": {
                        "type": "string",
                        "example": "REF001"
                    },
                    "shellId": {
                        "type": "integer",
                        "example": 123
                    },
                    "inventoryStatus": {
                        "type": "string",
                        "enum": [
                            "INSTOCK",
                            "LOWSTOCK",
                            "OUTOFSTOCK"
                        ],
                        "example": "INSTOCK"
                    },
                    "rating": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "utilisateur_demo"
                    },
                    "firstname": {
                        "type": "string",
                        "example": "Demo"
                    },
                    "email": {
                        "type": "string",
                        "example": "demo@example.com"
                    }
                },
                "type": "object"
            },
            "Wishlist": {
                "title": "Wishlist",
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Cart",
            "description": "Cart"
        },
        {
            "name": "Products",
            "description": "Products"
        },
        {
            "name": "Wishlist",
            "description": "Wishlist"
        }
    ]
}